[iOS] refactor: ImageUpdatable 프로토콜 적용
- image 정보들에 대한 id, image들을 가짐.
- image URL을 대입하면 해당하는 인덱스를 반환하는 메소드를 기본 구현 함.

#31
